fix 1.3

+++ AdminPageAuth.xaml.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.Linq;
 using System.Text;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
@@ -47,11 +48,25 @@ namespace Practice5
         {
             try
             {
-                object id = (dg_BD.SelectedItem as DataRowView).Row[0];
-                auth.DeleteQuery(Convert.ToInt32(id));
-                dg_BD.ItemsSource = auth.GetData();
-                dg_BD.Columns[0].Visibility = Visibility.Collapsed;
-
+                if (pole2.Text.Length < 5 && pole2.Text.Length > 15 != (IsValidInput(pole2.Text)) || !IsValidInput(pole2.Text))
+                {
+                    MessageBox.Show("Логин и пароль должны содержать от 5 до 15 букв и цифр!", "Ошибка", MessageBoxButton.OK, MessageBoxImage.Warning);
+                    return;
+                }
+                else
+                { 
+                    if (string.IsNullOrEmpty(pole1.Text) || string.IsNullOrEmpty(pole2.Text))
+                    {
+                        MessageBox.Show("Поля не имеют никакого внутри значения! Введите в поля данные");
+                    }
+                    else
+                    {
+                        object id = (dg_BD.SelectedItem as DataRowView).Row[0];
+                        auth.DeleteQuery(Convert.ToInt32(id));
+                        dg_BD.ItemsSource = auth.GetData();
+                        dg_BD.Columns[0].Visibility = Visibility.Collapsed;
+                    }
+                }
             }
             catch (Exception ex)
             {
@@ -61,21 +76,55 @@ namespace Practice5
 
         private void Add_Click(object sender, RoutedEventArgs e)
         {
-            auth.InsertQuery(pole1.Text, pole2.Text);
-            dg_BD.ItemsSource = auth.GetData();
-            dg_BD.Columns[0].Visibility = Visibility.Collapsed;
+            if (pole2.Text.Length < 5 && pole2.Text.Length > 15 != (IsValidInput(pole2.Text)) || !IsValidInput(pole2.Text))
+            {
+                MessageBox.Show("Логин и пароль должны содержать от 5 до 15 букв и цифр!", "Ошибка", MessageBoxButton.OK, MessageBoxImage.Warning);
+                return;
+            }
+            else
+            {
+                if (string.IsNullOrEmpty(pole1.Text) || string.IsNullOrEmpty(pole2.Text))
+                {
+                    MessageBox.Show("Поля не имеют никакого внутри значения! Введите в поля данные");
+                }
+                else
+                {
+                    auth.InsertQuery(pole1.Text, pole2.Text);
+                    dg_BD.ItemsSource = auth.GetData();
+                    dg_BD.Columns[0].Visibility = Visibility.Collapsed;
+                }
+            }
+        }
 
+        private bool IsValidInput(string input)
+        {
+            Regex regex = new Regex("^[a-zA-Z0-9]*$");
+            return regex.IsMatch(input);
         }
 
         private void Edit_Click(object sender, RoutedEventArgs e)
         {
             try
             {
-                object id = (dg_BD.SelectedItem as DataRowView).Row[0];
-                auth.UpdateQuery(pole1.Text, pole2.Text, Convert.ToInt32(id));
-                dg_BD.ItemsSource = auth.GetData();
-                dg_BD.Columns[0].Visibility = Visibility.Collapsed;
-
+                if (pole2.Text.Length < 5 && pole2.Text.Length > 15 != (IsValidInput(pole2.Text)) || !IsValidInput(pole2.Text))
+                {
+                    MessageBox.Show("Логин и пароль должны содержать от 5 до 15 букв и цифр!", "Ошибка", MessageBoxButton.OK, MessageBoxImage.Warning);
+                    return;
+                }
+                else
+                {
+                    if (string.IsNullOrEmpty(pole1.Text) || string.IsNullOrEmpty(pole2.Text))
+                    {
+                        MessageBox.Show("Поля не имеют никакого внутри значения! Введите в поля данные");
+                    }
+                    else
+                    {
+                        object id = (dg_BD.SelectedItem as DataRowView).Row[0];
+                        auth.UpdateQuery(pole1.Text, pole2.Text, Convert.ToInt32(id));
+                        dg_BD.ItemsSource = auth.GetData();
+                        dg_BD.Columns[0].Visibility = Visibility.Collapsed;
+                    }
+                }
             }
             catch
             {
